collect metrics (#7)
* i think this is wrong

* input copyimg, spawn vm

* dir, 24

* rm

* cp

* cores

* spanw dpdk

* rm

* fix allocatio

* 32 socks

* disable nat, reorg ip

* fp core max

* spanw dpdk vm

* fix sock

* rm

* log

* vtap

* fix collector

* log

* 4096 queue

* more hp

* iperf client logging

* py coll

* new coll

* m

* clear outdir

* trim iperf export

* commands

* comamnf

* just console

* comment out tap

* new consoel

* process

* 二四

+++ setup/copy_img.sh
@@ -1,10 +1,19 @@
 #! /bin/bash
 
-DEST_DIR=/proj/faasnetworkstack-PG0/testing
+if [ "$#" -ne 2 ]; then
+    echo "Usage: $0 <num_vms> <dest_dir>"
+    echo "  num_vms: number of VMs to copy"
+    echo "  dest_dir: destination directory"
+    exit 1
+fi
+
+NUM_VMS=$1
+# e.g. /proj/faasnetworkstack-PG0/testing
+DEST_DIR=$2
 
 sudo rm -rf $DEST_DIR/kernels/* $DEST_DIR/images/*
 sudo mkdir -p $DEST_DIR/kernels $DEST_DIR/images
-for i in {0..10}; do
+for ((i=0; i<NUM_VMS; i++)); do
   (
     sudo cp /tmp/vmlinux.bin $DEST_DIR/kernels/vm$i-kernel.bin
     sudo cp /tmp/noble-server-cloudimg-amd64.raw $DEST_DIR/images/vm$i-img.raw
